Fix loo calendar button and refactor calendar modal component
