+++ src/components/elements/calendar/CalendarDate.tsx
@@ -1,17 +1,15 @@
 "use client";
 
 import { format, isSameMonth, isToday } from "date-fns";
-import { MotionConfig } from "framer-motion";
 import { useState } from "react";
-import { HiX } from "react-icons/hi";
 import useMeasure from "react-use-measure";
 
-import { AnimationWrapper, Event, Modal, ResizablePanel, animations } from "@/components";
+import { AnimationWrapper, Event, animations } from "@/components";
 import { useUser } from "@/hooks";
 import { cn } from "@/lib";
 import { EventData } from "@/types";
 
-import { EventCreationTabs } from ".";
+import { CalendarEventCreationModal } from ".";
 
 interface Props {
   events: EventData[];
@@ -20,8 +18,6 @@ interface Props {
   isFetched: boolean;
   isAnimating: boolean;
 }
-
-const transition = { type: "ease", ease: "easeInOut", duration: 1 };
 
 export const CalendarDate = ({ events, date, month }: Props) => {
   const [ref, bounds] = useMeasure();
@@ -32,17 +28,11 @@ export const CalendarDate = ({ events, date, month }: Props) => {
     setIsModalOpen(true);
   };
 
-  const closeModal = () => {
-    setIsModalOpen(false);
-  };
-
   return (
-    <Modal
-      open={isModalOpen}
-      setOpen={setIsModalOpen}
-      maxWidth="2xl"
-      closeOnOverlayClick={false}
-      modalButton={
+    <CalendarEventCreationModal
+      isModalOpen={isModalOpen}
+      setIsModalOpen={setIsModalOpen}
+      button={
         <button
           ref={ref}
           className={cn(
@@ -92,26 +82,6 @@ export const CalendarDate = ({ events, date, month }: Props) => {
           </div>
         </button>
       }
-    >
-      <>
-        <MotionConfig transition={transition}>
-          <div className="relative overflow-hidden">
-            <ResizablePanel duration={transition.duration}>
-              <div className="flex flex-row items-center justify-between pt-6 px-7">
-                <p className="text-xl font-bold">Loo trenn</p>
-                <div role="button" tabIndex={0} onClick={closeModal}>
-                  <AnimationWrapper key="sub-modal-x-icon" variants={animations.rotate360}>
-                    <HiX className="w-8 h-8 fill-stone-700 hover:fill-stone-800" />
-                  </AnimationWrapper>
-                </div>
-              </div>
-              <div className="p-10">
-                <EventCreationTabs />
-              </div>
-            </ResizablePanel>
-          </div>
-        </MotionConfig>
-      </>
-    </Modal>
+    />
   );
 };
