+++ src/components/elements/calendar/event-creation/index.ts
@@ -1,4 +1,5 @@
 export * from "./AllDayEventInputs";
 export * from "./EventCreationTabs";
 export * from "./CalendarEventTab";
+export * from "./CalendarEventCreationModal";
 export * from "./time-date-pickers";
